feat: move `Action` and `ActionResponse` into base
