+++ uplink/src/collector/simulator.rs
@@ -10,9 +10,8 @@ use std::collections::{BinaryHeap, HashMap};
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 use std::{cmp::Ordering, fs, io, sync::Arc};
 
-use crate::base::SimulatorConfig;
-use crate::base::{middleware::Action, Package, Stream};
-use crate::{ActionResponse, Payload};
+use crate::base::{Package, SimulatorConfig, Stream};
+use crate::{Action, ActionResponse, Payload};
 
 use rand::Rng;
